introduced color palette, pulled images, general cleanup
Wasn't happy working with images, so I found a nice rainbow color palette based on the nodeschool sticker orange. Rainbow because that's sort of what I feel like SF's colors are, when we aren't stuck with sports teams. Cleaned up some of the markup and text, put in a section with the new event info.